+++ src/Equinor.Procosys.Preservation.Command/TagCommands/SetStep/SetStepCommandValidator.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using Equinor.Procosys.Preservation.Command.Validators.Step;
 using Equinor.Procosys.Preservation.Command.Validators.Tag;
-using Equinor.Procosys.Preservation.Command.Validators.Project;
+using Equinor.Procosys.Preservation.Command.Validators.ProjectValidators;
 using FluentValidation;
 
 namespace Equinor.Procosys.Preservation.Command.TagCommands.SetStep
