+++ src/Equinor.Procosys.Preservation.Command/TagCommands/StartPreservation/StartPreservationCommandValidator.cs
@@ -4,7 +4,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Equinor.Procosys.Preservation.Command.Validators.Tag;
 using Equinor.Procosys.Preservation.Domain.AggregateModels.ProjectAggregate;
-using Equinor.Procosys.Preservation.Command.Validators.Project;
+using Equinor.Procosys.Preservation.Command.Validators.ProjectValidators;
 using FluentValidation;
 
 namespace Equinor.Procosys.Preservation.Command.TagCommands.StartPreservation
